feat: return organization details with list and create project (#67)
For a free organization feature, new projects should be created in the
default organization and not under a personal account. This PR adds
additional logic to resolve the user's organization for `list_projects`
and `create_project` tools


Resolution: 

1. Fetch user details and verify if the project can be created under a
personal/individual account. If yes, proceed with the tool execution
2. If not, the user should have a free organization. Fetch and filter
organizations managed by `console`
3. If the filter results in only one organization, execute the tool
under this organization scope
4. If the filter returns multiple orgs, throw error requesting
user/agent for an `org_id`


cc: @pffigueiredo @davidgomes