fix: issue with using org api keys while trying to identify users (#70)
We recently made a change to fetch the user details for `list_projects`
and `create_project` tools to identify if the user has an org-only
policy. However, when the server is started in stdio with the org API
key, the `getCurrentUser` API fails.

This PR makes changes to identify an <ins>_account_</ins> instead of
just _user_, the `account.id` is preserved and passed to tool handlers
in the context. If an account is an ORG, we do not call `getCurrentUser`
and instead `getOrganization(account.id)`

The `listProjects` is used in multiple tools, and I had to pass the
context around, but it was necessary to make the `extra.account.id`
available to correctly fetch the projects of the current account.

> [!NOTE]
> Review the refactoring done in #69 before this PR

cc: @davidgomes